refactor(server): test fixtures
