+++ server/src/domain/auth/auth.service.spec.ts
@@ -12,8 +12,8 @@ import {
   newUserTokenRepositoryMock,
   sharedLinkStub,
   systemConfigStub,
-  userEntityStub,
-  userTokenEntityStub,
+  userStub,
+  userTokenStub,
 } from '@test';
 import { IncomingHttpHeaders } from 'http';
 import { generators, Issuer } from 'openid-client';
@@ -112,15 +112,15 @@ describe('AuthService', () => {
     });
 
     it('should successfully log the user in', async () => {
-      userMock.getByEmail.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.create.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.getByEmail.mockResolvedValue(userStub.user1);
+      userTokenMock.create.mockResolvedValue(userTokenStub.userToken);
       await expect(sut.login(fixtures.login, loginDetails)).resolves.toEqual(loginResponseStub.user1password);
       expect(userMock.getByEmail).toHaveBeenCalledTimes(1);
     });
 
     it('should generate the cookie headers (insecure)', async () => {
-      userMock.getByEmail.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.create.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.getByEmail.mockResolvedValue(userStub.user1);
+      userTokenMock.create.mockResolvedValue(userTokenStub.userToken);
       await expect(
         sut.login(fixtures.login, {
           clientIp: '127.0.0.1',
@@ -246,10 +246,10 @@ describe('AuthService', () => {
     });
 
     it('should validate using authorization header', async () => {
-      userMock.get.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.getByToken.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.get.mockResolvedValue(userStub.user1);
+      userTokenMock.getByToken.mockResolvedValue(userTokenStub.userToken);
       const client = { request: { headers: { authorization: 'Bearer auth_token' } } };
-      await expect(sut.validate((client as Socket).request.headers, {})).resolves.toEqual(userEntityStub.user1);
+      await expect(sut.validate((client as Socket).request.headers, {})).resolves.toEqual(userStub.user1);
     });
   });
 
@@ -275,7 +275,7 @@ describe('AuthService', () => {
 
     it('should accept a base64url key', async () => {
       shareMock.getByKey.mockResolvedValue(sharedLinkStub.valid);
-      userMock.get.mockResolvedValue(userEntityStub.admin);
+      userMock.get.mockResolvedValue(userStub.admin);
       const headers: IncomingHttpHeaders = { 'x-immich-share-key': sharedLinkStub.valid.key.toString('base64url') };
       await expect(sut.validate(headers, {})).resolves.toEqual(authStub.adminSharedLink);
       expect(shareMock.getByKey).toHaveBeenCalledWith(sharedLinkStub.valid.key);
@@ -283,7 +283,7 @@ describe('AuthService', () => {
 
     it('should accept a hex key', async () => {
       shareMock.getByKey.mockResolvedValue(sharedLinkStub.valid);
-      userMock.get.mockResolvedValue(userEntityStub.admin);
+      userMock.get.mockResolvedValue(userStub.admin);
       const headers: IncomingHttpHeaders = { 'x-immich-share-key': sharedLinkStub.valid.key.toString('hex') };
       await expect(sut.validate(headers, {})).resolves.toEqual(authStub.adminSharedLink);
       expect(shareMock.getByKey).toHaveBeenCalledWith(sharedLinkStub.valid.key);
@@ -298,16 +298,16 @@ describe('AuthService', () => {
     });
 
     it('should return an auth dto', async () => {
-      userTokenMock.getByToken.mockResolvedValue(userTokenEntityStub.userToken);
+      userTokenMock.getByToken.mockResolvedValue(userTokenStub.userToken);
       const headers: IncomingHttpHeaders = { cookie: 'immich_access_token=auth_token' };
-      await expect(sut.validate(headers, {})).resolves.toEqual(userEntityStub.user1);
+      await expect(sut.validate(headers, {})).resolves.toEqual(userStub.user1);
     });
 
     it('should update when access time exceeds an hour', async () => {
-      userTokenMock.getByToken.mockResolvedValue(userTokenEntityStub.inactiveToken);
-      userTokenMock.save.mockResolvedValue(userTokenEntityStub.userToken);
+      userTokenMock.getByToken.mockResolvedValue(userTokenStub.inactiveToken);
+      userTokenMock.save.mockResolvedValue(userTokenStub.userToken);
       const headers: IncomingHttpHeaders = { cookie: 'immich_access_token=auth_token' };
-      await expect(sut.validate(headers, {})).resolves.toEqual(userEntityStub.user1);
+      await expect(sut.validate(headers, {})).resolves.toEqual(userStub.user1);
       expect(userTokenMock.save.mock.calls[0][0]).toMatchObject({
         id: 'not_active',
         token: 'auth_token',
@@ -338,7 +338,7 @@ describe('AuthService', () => {
 
   describe('getDevices', () => {
     it('should get the devices', async () => {
-      userTokenMock.getAll.mockResolvedValue([userTokenEntityStub.userToken, userTokenEntityStub.inactiveToken]);
+      userTokenMock.getAll.mockResolvedValue([userTokenStub.userToken, userTokenStub.inactiveToken]);
       await expect(sut.getDevices(authStub.user1)).resolves.toEqual([
         {
           createdAt: '2021-01-01T00:00:00.000Z',
@@ -364,7 +364,7 @@ describe('AuthService', () => {
 
   describe('logoutDevices', () => {
     it('should logout all devices', async () => {
-      userTokenMock.getAll.mockResolvedValue([userTokenEntityStub.inactiveToken, userTokenEntityStub.userToken]);
+      userTokenMock.getAll.mockResolvedValue([userTokenStub.inactiveToken, userTokenStub.userToken]);
 
       await sut.logoutDevices(authStub.user1);
 
@@ -429,24 +429,24 @@ describe('AuthService', () => {
 
     it('should link an existing user', async () => {
       configMock.load.mockResolvedValue(systemConfigStub.noAutoRegister);
-      userMock.getByEmail.mockResolvedValue(userEntityStub.user1);
-      userMock.update.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.create.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.getByEmail.mockResolvedValue(userStub.user1);
+      userMock.update.mockResolvedValue(userStub.user1);
+      userTokenMock.create.mockResolvedValue(userTokenStub.userToken);
 
       await expect(sut.callback({ url: 'http://immich/auth/login?code=abc123' }, loginDetails)).resolves.toEqual(
         loginResponseStub.user1oauth,
       );
 
       expect(userMock.getByEmail).toHaveBeenCalledTimes(1);
-      expect(userMock.update).toHaveBeenCalledWith(userEntityStub.user1.id, { oauthId: sub });
+      expect(userMock.update).toHaveBeenCalledWith(userStub.user1.id, { oauthId: sub });
     });
 
     it('should allow auto registering by default', async () => {
       configMock.load.mockResolvedValue(systemConfigStub.enabled);
       userMock.getByEmail.mockResolvedValue(null);
-      userMock.getAdmin.mockResolvedValue(userEntityStub.user1);
-      userMock.create.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.create.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.getAdmin.mockResolvedValue(userStub.user1);
+      userMock.create.mockResolvedValue(userStub.user1);
+      userTokenMock.create.mockResolvedValue(userTokenStub.userToken);
 
       await expect(sut.callback({ url: 'http://immich/auth/login?code=abc123' }, loginDetails)).resolves.toEqual(
         loginResponseStub.user1oauth,
@@ -458,8 +458,8 @@ describe('AuthService', () => {
 
     it('should use the mobile redirect override', async () => {
       configMock.load.mockResolvedValue(systemConfigStub.override);
-      userMock.getByOAuthId.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.create.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.getByOAuthId.mockResolvedValue(userStub.user1);
+      userTokenMock.create.mockResolvedValue(userTokenStub.userToken);
 
       await sut.callback({ url: `app.immich:/?code=abc123` }, loginDetails);
 
@@ -468,8 +468,8 @@ describe('AuthService', () => {
 
     it('should use the mobile redirect override for ios urls with multiple slashes', async () => {
       configMock.load.mockResolvedValue(systemConfigStub.override);
-      userMock.getByOAuthId.mockResolvedValue(userEntityStub.user1);
-      userTokenMock.create.mockResolvedValue(userTokenEntityStub.userToken);
+      userMock.getByOAuthId.mockResolvedValue(userStub.user1);
+      userTokenMock.create.mockResolvedValue(userTokenStub.userToken);
 
       await sut.callback({ url: `app.immich:///?code=abc123` }, loginDetails);
 
@@ -480,7 +480,7 @@ describe('AuthService', () => {
   describe('link', () => {
     it('should link an account', async () => {
       configMock.load.mockResolvedValue(systemConfigStub.enabled);
-      userMock.update.mockResolvedValue(userEntityStub.user1);
+      userMock.update.mockResolvedValue(userStub.user1);
 
       await sut.link(authStub.user1, { url: 'http://immich/user-settings?code=abc123' });
 
@@ -502,7 +502,7 @@ describe('AuthService', () => {
   describe('unlink', () => {
     it('should unlink an account', async () => {
       configMock.load.mockResolvedValue(systemConfigStub.enabled);
-      userMock.update.mockResolvedValue(userEntityStub.user1);
+      userMock.update.mockResolvedValue(userStub.user1);
 
       await sut.unlink(authStub.user1);
