+++ server/src/domain/person/person.service.spec.ts
@@ -1,6 +1,6 @@
 import { BadRequestException, NotFoundException } from '@nestjs/common';
 import {
-  assetEntityStub,
+  assetStub,
   authStub,
   faceStub,
   newJobRepositoryMock,
@@ -112,7 +112,7 @@ describe(PersonService.name, () => {
 
   describe('getAssets', () => {
     it("should return a person's assets", async () => {
-      personMock.getAssets.mockResolvedValue([assetEntityStub.image, assetEntityStub.video]);
+      personMock.getAssets.mockResolvedValue([assetStub.image, assetStub.video]);
       await sut.getAssets(authStub.admin, 'person-1');
       expect(personMock.getAssets).toHaveBeenCalledWith('admin_id', 'person-1');
     });
@@ -130,7 +130,7 @@ describe(PersonService.name, () => {
     it("should update a person's name", async () => {
       personMock.getById.mockResolvedValue(personStub.noName);
       personMock.update.mockResolvedValue(personStub.withName);
-      personMock.getAssets.mockResolvedValue([assetEntityStub.image]);
+      personMock.getAssets.mockResolvedValue([assetStub.image]);
 
       await expect(sut.update(authStub.admin, 'person-1', { name: 'Person 1' })).resolves.toEqual(responseDto);
 
@@ -138,14 +138,14 @@ describe(PersonService.name, () => {
       expect(personMock.update).toHaveBeenCalledWith({ id: 'person-1', name: 'Person 1' });
       expect(jobMock.queue).toHaveBeenCalledWith({
         name: JobName.SEARCH_INDEX_ASSET,
-        data: { ids: [assetEntityStub.image.id] },
+        data: { ids: [assetStub.image.id] },
       });
     });
 
     it('should update a person visibility', async () => {
       personMock.getById.mockResolvedValue(personStub.hidden);
       personMock.update.mockResolvedValue(personStub.withName);
-      personMock.getAssets.mockResolvedValue([assetEntityStub.image]);
+      personMock.getAssets.mockResolvedValue([assetStub.image]);
 
       await expect(sut.update(authStub.admin, 'person-1', { isHidden: false })).resolves.toEqual(responseDto);
 
@@ -153,7 +153,7 @@ describe(PersonService.name, () => {
       expect(personMock.update).toHaveBeenCalledWith({ id: 'person-1', isHidden: false });
       expect(jobMock.queue).toHaveBeenCalledWith({
         name: JobName.SEARCH_INDEX_ASSET,
-        data: { ids: [assetEntityStub.image.id] },
+        data: { ids: [assetStub.image.id] },
       });
     });
 
@@ -239,7 +239,7 @@ describe(PersonService.name, () => {
     it('should delete conflicting faces before merging', async () => {
       personMock.getById.mockResolvedValue(personStub.primaryPerson);
       personMock.getById.mockResolvedValue(personStub.mergePerson);
-      personMock.prepareReassignFaces.mockResolvedValue([assetEntityStub.image.id]);
+      personMock.prepareReassignFaces.mockResolvedValue([assetStub.image.id]);
 
       await expect(sut.mergePerson(authStub.admin, 'person-1', { ids: ['person-2'] })).resolves.toEqual([
         { id: 'person-2', success: true },
@@ -252,7 +252,7 @@ describe(PersonService.name, () => {
 
       expect(jobMock.queue).toHaveBeenCalledWith({
         name: JobName.SEARCH_REMOVE_FACE,
-        data: { assetId: assetEntityStub.image.id, personId: personStub.mergePerson.id },
+        data: { assetId: assetStub.image.id, personId: personStub.mergePerson.id },
       });
     });
 
@@ -282,7 +282,7 @@ describe(PersonService.name, () => {
     it('should handle an error reassigning faces', async () => {
       personMock.getById.mockResolvedValue(personStub.primaryPerson);
       personMock.getById.mockResolvedValue(personStub.mergePerson);
-      personMock.prepareReassignFaces.mockResolvedValue([assetEntityStub.image.id]);
+      personMock.prepareReassignFaces.mockResolvedValue([assetStub.image.id]);
       personMock.reassignFaces.mockRejectedValue(new Error('update failed'));
 
       await expect(sut.mergePerson(authStub.admin, 'person-1', { ids: ['person-2'] })).resolves.toEqual([
