+++ server/test/fixtures/index.ts
@@ -1,0 +1,15 @@
+export * from './album.stub';
+export * from './api-key.stub';
+export * from './asset.stub';
+export * from './auth.stub';
+export * from './face.stub';
+export * from './file.stub';
+export * from './media.stub';
+export * from './partner.stub';
+export * from './person.stub';
+export * from './search.stub';
+export * from './shared-link.stub';
+export * from './system-config.stub';
+export * from './tag.stub';
+export * from './user-token.stub';
+export * from './user.stub';
